+++ runtime-metamodel/src/test/java/org/eclipse/dataspaceconnector/tooling/autodoc/domain/EdcModuleTest.java
@@ -16,6 +16,10 @@ package org.eclipse.dataspaceconnector.tooling.autodoc.domain;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.eclipse.dataspaceconnector.runtime.metamodel.domain.ConfigurationSetting;
+import org.eclipse.dataspaceconnector.runtime.metamodel.domain.EdcModule;
+import org.eclipse.dataspaceconnector.runtime.metamodel.domain.Service;
+import org.eclipse.dataspaceconnector.runtime.metamodel.domain.ServiceReference;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
